+++ decoder/packet_ethernet.h
@@ -59,7 +59,7 @@ uint32_t get_eth802_oui(struct snap_info *snap);
 char *get_ethernet_type(uint16_t ethertype);
 
 /* Should be internal to the decoder */
-bool handle_ethernet(unsigned char *buffer, size_t n, struct packet *p);
+bool handle_ethernet(unsigned char *buffer, int n, struct packet *p);
 void free_ethernet802_3_frame(struct eth_info *eth);
 
 #endif
