+++ decoder/packet_udp.c
@@ -15,7 +15,7 @@
  * +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
  *
  */
-packet_error handle_udp(unsigned char *buffer, size_t n, struct udp_info *info)
+packet_error handle_udp(unsigned char *buffer, int n, struct udp_info *info)
 {
     if (n < UDP_HDR_LEN) return UDP_ERR;
