+++ decoder/packet_udp.h
@@ -17,6 +17,6 @@ struct udp_info {
 };
 
 /* internal to the decoder */
-packet_error handle_udp(unsigned char *buffer, size_t n, struct udp_info *info);
+packet_error handle_udp(unsigned char *buffer, int n, struct udp_info *info);
 
 #endif
